Further simplification of the BaseNeighbourhoodProcesing class
Moved BaseNeighbourhoodProcessing unit tests.
Removed slicing over time and suport for multi-time cubes
