+++ improver_tests/nbhood/circular_kernel/test_GeneratePercentilesFromACircularNeighbourhood.py
@@ -330,9 +330,9 @@ class Test_pad_and_unpad_cube(IrisTest):
         self.assertArrayAlmostEqual(result.data, expected)
 
 
-class Test_run(IrisTest):
+class Test_process(IrisTest):
 
-    """Test the run method within the plugin to calculate percentile values
+    """Test the process method within the plugin to calculate percentile values
     from a neighbourhood."""
 
     def setUp(self):
@@ -346,7 +346,7 @@ class Test_run(IrisTest):
 
         self.cube.data[2, 2] = 0
         radius = 4000.0
-        result = GeneratePercentilesFromANeighbourhood(radius).run(self.cube)
+        result = GeneratePercentilesFromANeighbourhood(radius).process(self.cube)
         self.assertIsInstance(result, Cube)
 
     def test_single_point(self):
@@ -385,7 +385,7 @@ class Test_run(IrisTest):
         radius = 2000.0
         result = GeneratePercentilesFromANeighbourhood(
             radius, percentiles=percentiles
-        ).run(cube)
+        ).process(cube)
         self.assertArrayAlmostEqual(result.data, expected)
 
     def test_multi_point_multitimes(self):
@@ -459,7 +459,7 @@ class Test_run(IrisTest):
         radius = 2000.0
         result = GeneratePercentilesFromANeighbourhood(
             radius, percentiles=percentiles
-        ).run(cube)
+        ).process(cube)
         self.assertArrayAlmostEqual(result.data, expected)
 
     def test_single_point_lat_long(self):
@@ -472,7 +472,7 @@ class Test_run(IrisTest):
         msg = "Unable to convert from"
         radius = 6000.0
         with self.assertRaisesRegex(ValueError, msg):
-            GeneratePercentilesFromANeighbourhood(radius).run(cube)
+            GeneratePercentilesFromANeighbourhood(radius).process(cube)
 
     def test_single_point_masked_to_null(self):
         """Test behaviour with a masked non-zero point.
@@ -514,7 +514,7 @@ class Test_run(IrisTest):
         radius = 2000.0
         result = GeneratePercentilesFromANeighbourhood(
             radius, percentiles=percentiles
-        ).run(self.cube)
+        ).process(self.cube)
         self.assertArrayAlmostEqual(result.data, expected)
 
     def test_single_point_masked_other_point(self):
@@ -555,7 +555,7 @@ class Test_run(IrisTest):
         radius = 2000.0
         result = GeneratePercentilesFromANeighbourhood(
             radius, percentiles=percentiles
-        ).run(self.cube)
+        ).process(self.cube)
         self.assertArrayAlmostEqual(result.data, expected)
 
     def test_single_point_low_percentiles(self):
@@ -592,7 +592,7 @@ class Test_run(IrisTest):
         radius = 2000.0
         result = GeneratePercentilesFromANeighbourhood(
             radius, percentiles=percentiles
-        ).run(self.cube)
+        ).process(self.cube)
         self.assertArrayAlmostEqual(result.data, expected)
 
     def test_point_pair(self):
@@ -630,7 +630,7 @@ class Test_run(IrisTest):
         radius = 2000.0
         result = GeneratePercentilesFromANeighbourhood(
             radius, percentiles=percentiles
-        ).run(self.cube)
+        ).process(self.cube)
         self.assertArrayAlmostEqual(result.data, expected)
 
     def test_number_of_percentiles_equals_number_of_points(self):
@@ -694,7 +694,7 @@ class Test_run(IrisTest):
         radius = 2000.0
         result = GeneratePercentilesFromANeighbourhood(
             radius, percentiles=percentiles
-        ).run(cube)
+        ).process(cube)
         self.assertArrayAlmostEqual(result.data, expected)
 
     def test_number_of_points_half_of_number_of_percentiles(self):
@@ -803,7 +803,7 @@ class Test_run(IrisTest):
         radius = 2000.0
         result = GeneratePercentilesFromANeighbourhood(
             radius, percentiles=percentiles
-        ).run(cube)
+        ).process(cube)
         self.assertArrayAlmostEqual(result.data, expected)
 
     def test_circle_bigger_than_domain(self):
@@ -814,7 +814,7 @@ class Test_run(IrisTest):
         radius = 50000.0
         msg = "Distance of {}m exceeds max domain distance".format(radius)
         with self.assertRaisesRegex(ValueError, msg):
-            GeneratePercentilesFromANeighbourhood(radius).run(self.cube)
+            GeneratePercentilesFromANeighbourhood(radius).process(self.cube)
 
     def test_mask_cube(self):
         """Test that a NotImplementedError is raised if a mask cube is passed
@@ -825,7 +825,7 @@ class Test_run(IrisTest):
         radius = 4000.0
         msg = "The use of a mask cube with a circular kernel is " "not yet implemented."
         with self.assertRaisesRegex(NotImplementedError, msg):
-            GeneratePercentilesFromANeighbourhood(radius).run(
+            GeneratePercentilesFromANeighbourhood(radius).process(
                 self.cube, mask_cube=self.cube
             )
