+++ improver_tests/nbhood/square_kernel/test_SquareNeighbourhood.py
@@ -55,9 +55,9 @@ class Test__init__(IrisTest):
             NeighbourhoodProcessing("square", 2000, sum_or_fraction=sum_or_fraction)
 
 
-class Test_run(IrisTest):
+class Test_process(IrisTest):
 
-    """Test the run method on the NeighbourhoodProcessing class."""
+    """Test the process method on the NeighbourhoodProcessing class."""
 
     RADIUS = 2500
 
@@ -78,7 +78,7 @@ class Test_run(IrisTest):
         )
 
     def test_basic_re_mask_true(self):
-        """Test that a cube with correct data is produced by the run method
+        """Test that a cube with correct data is produced by the process method
         when re-masking is applied."""
         expected_array = np.array(
             [
@@ -89,7 +89,7 @@ class Test_run(IrisTest):
                 [1.0, 1.0, 1.0, 1.0, 1.0],
             ]
         )
-        result = NeighbourhoodProcessing("square", self.RADIUS).run(self.cube)
+        result = NeighbourhoodProcessing("square", self.RADIUS).process(self.cube)
         self.assertIsInstance(result, Cube)
         self.assertArrayAlmostEqual(result.data, expected_array)
 
@@ -110,11 +110,11 @@ class Test_run(IrisTest):
         self.cube.coord("projection_x_coordinate").points = coord_points_x
         self.cube.coord("projection_y_coordinate").points = coord_points_y
 
-        result = NeighbourhoodProcessing("square", self.RADIUS).run(self.cube)
+        result = NeighbourhoodProcessing("square", self.RADIUS).process(self.cube)
         self.assertArrayAlmostEqual(result.data, expected_array)
 
     def test_basic_re_mask_false(self):
-        """Test that a cube with correct data is produced by the run method."""
+        """Test that a cube with correct data is produced by the process method."""
         expected_array = np.array(
             [
                 [1.0, 1.0, 1.0, 1.0, 1.0],
@@ -124,13 +124,13 @@ class Test_run(IrisTest):
                 [1.0, 1.0, 1.0, 1.0, 1.0],
             ]
         )
-        result = NeighbourhoodProcessing("square", self.RADIUS, re_mask=False).run(
+        result = NeighbourhoodProcessing("square", self.RADIUS, re_mask=False).process(
             self.cube,
         )
         self.assertArrayAlmostEqual(result.data, expected_array)
 
     def test_masked_array_re_mask_true(self):
-        """Test that the run method produces a cube with correct data when a
+        """Test that the process method produces a cube with correct data when a
         cube containing masked data is passed in and re-masking is applied."""
         self.cube.data = np.array(
             [
@@ -169,12 +169,12 @@ class Test_run(IrisTest):
             ]
         )
         self.cube.data = np.ma.masked_where(mask == 0, self.cube.data)
-        result = NeighbourhoodProcessing("square", self.RADIUS).run(self.cube)
+        result = NeighbourhoodProcessing("square", self.RADIUS).process(self.cube)
         self.assertArrayAlmostEqual(result.data.data, expected_array)
         self.assertArrayAlmostEqual(result.data.mask, expected_mask_array)
 
     def test_masked_array_re_mask_false(self):
-        """Test that the run method produces a cube with correct data when a
+        """Test that the process method produces a cube with correct data when a
            cube containing masked data is passed in."""
         self.cube.data = np.array(
             [
@@ -204,46 +204,13 @@ class Test_run(IrisTest):
             ]
         )
         self.cube.data = np.ma.masked_where(mask == 0, self.cube.data)
-        result = NeighbourhoodProcessing("square", self.RADIUS, re_mask=False).run(
-            self.cube,
-        )
-        self.assertArrayAlmostEqual(result.data, expected_array)
-
-    def test_nan_array_re_mask_true(self):
-        """Test that an array containing nans is handled correctly when
-        re-masking is applied."""
-        expected_array = np.array(
-            [
-                [np.nan, 1.0, 1.0, 1.0, 1.0],
-                [1.0, 0.8750000, 0.88888889, 0.88888889, 1.0],
-                [1.0, 0.88888889, 0.88888889, 0.88888889, 1.0],
-                [1.0, 0.88888889, 0.88888889, 0.88888889, 1.0],
-                [1.0, 1.0, 1.0, 1.0, 1.0],
-            ]
-        )
-        self.cube.data[0, 0] = np.nan
-        result = NeighbourhoodProcessing("square", self.RADIUS).run(self.cube,)
-        self.assertArrayAlmostEqual(result.data, expected_array)
-
-    def test_nan_array_re_mask_false(self):
-        """Test that an array containing nans is handled correctly."""
-        expected_array = np.array(
-            [
-                [np.nan, 1.0, 1.0, 1.0, 1.0],
-                [1.0, 0.8750000, 0.88888889, 0.88888889, 1.0],
-                [1.0, 0.88888889, 0.88888889, 0.88888889, 1.0],
-                [1.0, 0.88888889, 0.88888889, 0.88888889, 1.0],
-                [1.0, 1.0, 1.0, 1.0, 1.0],
-            ]
-        )
-        self.cube.data[0, 0] = np.nan
-        result = NeighbourhoodProcessing("square", self.RADIUS, re_mask=False).run(
+        result = NeighbourhoodProcessing("square", self.RADIUS, re_mask=False).process(
             self.cube,
         )
         self.assertArrayAlmostEqual(result.data, expected_array)
 
     def test_masked_array_with_nans_re_mask_true(self):
-        """Test that the run method produces a cube with correct data when a
+        """Test that the process method produces a cube with correct data when a
         cube containing masked nans is passed in and when re-masking is
         applied."""
         self.cube.data = np.array(
@@ -274,11 +241,11 @@ class Test_run(IrisTest):
             ]
         )
         self.cube.data = np.ma.masked_where(mask == 0, self.cube.data)
-        result = NeighbourhoodProcessing("square", self.RADIUS).run(self.cube,)
+        result = NeighbourhoodProcessing("square", self.RADIUS).process(self.cube,)
         self.assertArrayAlmostEqual(result.data, expected_array)
 
     def test_masked_array_with_nans_re_mask_false(self):
-        """Test that the run method produces a cube with correct data when a
+        """Test that the process method produces a cube with correct data when a
            cube containing masked nans is passed in."""
         self.cube.data = np.array(
             [
@@ -308,7 +275,7 @@ class Test_run(IrisTest):
             ]
         )
         self.cube.data = np.ma.masked_where(mask == 0, self.cube.data)
-        result = NeighbourhoodProcessing("square", self.RADIUS, re_mask=False).run(
+        result = NeighbourhoodProcessing("square", self.RADIUS, re_mask=False).process(
             self.cube,
         )
         self.assertArrayAlmostEqual(result.data, expected_array)
@@ -351,11 +318,11 @@ class Test_run(IrisTest):
                 [1.0 + 0.0j, 1.0 + 0.0j, 0.9 + 0.1j, 0.9 + 0.1j, 0.85 + 0.15j],
             ]
         )
-        result = NeighbourhoodProcessing("square", self.RADIUS).run(self.cube,)
+        result = NeighbourhoodProcessing("square", self.RADIUS).process(self.cube,)
         self.assertArrayAlmostEqual(result.data, expected_array)
 
     def test_multiple_realizations(self):
-        """Test that a cube with correct data is produced by the run method
+        """Test that a cube with correct data is produced by the process method
         when multiple realizations are supplied."""
         expected_1 = np.array(
             [
@@ -378,14 +345,14 @@ class Test_run(IrisTest):
 
         self.multi_realization_cube.data[0, 2, 2] = 0
         self.multi_realization_cube.data[1, 1, 2] = 0
-        result = NeighbourhoodProcessing("square", self.RADIUS).run(
+        result = NeighbourhoodProcessing("square", self.RADIUS).process(
             self.multi_realization_cube,
         )
         self.assertArrayAlmostEqual(result.data[0], expected_1)
         self.assertArrayAlmostEqual(result.data[1], expected_2)
 
     def test_multiple_realizations_with_mask(self):
-        """Test that the run method produces a cube with correct data when a
+        """Test that the process method produces a cube with correct data when a
         cube containing masked data at multiple realizations is passed in.
         Re-masking is disabled."""
 
@@ -445,47 +412,17 @@ class Test_run(IrisTest):
                 ],
             ]
         )
-        result = NeighbourhoodProcessing("square", self.RADIUS, re_mask=False).run(
+        result = NeighbourhoodProcessing("square", self.RADIUS, re_mask=False).process(
             self.multi_realization_cube,
         )
         self.assertArrayAlmostEqual(result.data, expected_array)
 
-    def test_multiple_realizations_nan(self):
-        """Test that a cube with correct data is produced by the run method
-        for multiple realizations and for when nans are present."""
-        expected_1 = np.array(
-            [
-                [np.nan, 0.8, 0.8333333, 0.8333333, 1.0],
-                [1.0, 0.75, 0.77777778, 0.77777778, 1.0],
-                [1.0, 0.77777778, 0.77777778, 0.77777778, 1.0],
-                [1.0, 0.88888889, 0.88888889, 0.88888889, 1.0],
-                [1.0, 1.0, 1.0, 1.0, 1.0],
-            ]
-        )
-        expected_2 = np.array(
-            [
-                [1.0, 1.0, 1.0, 1.0, 1.0],
-                [1.0, np.nan, 1.0, 1.0, 1.0],
-                [1.0, 1.0, 1.0, 1.0, 1.0],
-                [1.0, 1.0, 1.0, 1.0, 1.0],
-                [1.0, 1.0, 1.0, 1.0, 1.0],
-            ]
-        )
-        cube = self.multi_realization_cube
-        cube.data[0, 2, 2] = 0
-        cube.data[0, 1, 2] = 0
-        cube.data[0, 0, 0] = np.nan
-        cube.data[1, 1, 1] = np.nan
-        result = NeighbourhoodProcessing("square", self.RADIUS).run(cube)
-        self.assertArrayAlmostEqual(result.data[0], expected_1)
-        self.assertArrayAlmostEqual(result.data[1], expected_2)
-
     def test_metadata(self):
-        """Test that a cube with correct metadata is produced by the run
+        """Test that a cube with correct metadata is produced by the process
         method."""
         self.cube.attributes = {"Conventions": "CF-1.5"}
         self.cube.add_cell_method(CellMethod("mean", coords="time"))
-        result = NeighbourhoodProcessing("square", self.RADIUS).run(self.cube,)
+        result = NeighbourhoodProcessing("square", self.RADIUS).process(self.cube,)
         self.assertTupleEqual(result.cell_methods, self.cube.cell_methods)
         self.assertDictEqual(result.attributes, self.cube.attributes)
